DataService/requirements.txt
azure-functions==1.13.3
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
idna==3.4
pymongo==4.3.3
requests==2.28.2
urllib3==1.26.15
